fix: stop dashboard sse refresh flush churn
Keep dashboard freshness probes on no-flush summary and rollup reads.

Track pending rollup signatures and reuse rebuilt SSE freshness.

Split dashboard snapshot regressions into a dedicated test module.
