+++ docs/specs/README.md
@@ -109,7 +109,7 @@
 |  2c3ep | Release 原生 ARM 镜像发布与双架构 manifest             | 已完成                | `2c3ep-release-native-arm-images/SPEC.md`                      | 2026-04-04 | PR #207 merge-ready; release now publishes `linux/amd64` + `linux/arm64` manifests, with `linux/arm64` built on `ubuntu-24.04-arm` and review-loop clear                                                                                                                                                         |
 |  5y2g4 | Upstream/Base URL 路径前缀兼容增强                     | 已完成（merge-ready） | `5y2g4-upstream-base-url-path-prefix-compat/SPEC.md`           | 2026-04-04 | PR #208: shared URL helper, prefixed MCP/HTTP route coverage, green checks, and review-loop convergence are complete                                                                                                                                                                                             |
 |  kmmtg | Release 发布后回写 PR 评论                             | 已完成                | `kmmtg-release-pr-comment-upsert/SPEC.md`                      | 2026-04-06 | release workflow now upserts a marker-based bot comment onto the merged PR with release link, channel, version, and GHCR tags                                                                                                                                                                                    |
-|  66t8u | Admin 仪表盘总览加载去重与轻量快照                     | 已实现（待审查）      | `66t8u-admin-dashboard-overview-performance/SPEC.md`           | 2026-06-23 | dashboard overview keeps the shared lightweight payload contract and now emits default structured perf events for overview + shared-snapshot cache-hit/rebuild paths                                                                                                                                             |
+|  66t8u | Admin 仪表盘总览加载去重与轻量快照                     | 已实现（待审查）      | `66t8u-admin-dashboard-overview-performance/SPEC.md`           | 2026-06-24 | dashboard overview freshness now stays on a no-flush shared-snapshot probe with pending rollup signatures, emits rebuilt SSE freshness, and keeps the structured perf evidence path intact                                                                                                                       |
 |  h2698 | Admin 仪表盘 49 小时整点请求图表                       | 已完成                | `h2698-admin-dashboard-hourly-request-charts/SPEC.md`          | 2026-04-17 | hourly window semantics now show 25 groups over the last 24 hours with the current server-time hour included; yesterday delta keeps comparing against the same hour bucket while labels remain in local time                                                                                                     |
 |  ev4td | 管理台近期请求文案纠正与性能重构                       | 已实现（待审查）      | `ev4td-admin-recent-requests-performance-copy/SPEC.md`         | 2026-06-23 | admin recent requests list/catalog still use cursor + retention-aware copy, and now emit default structured perf events for global/token-scoped heavy reads                                                                                                                                                      |
 |  g9ku2 | Web 404 fallback 主题一致性修复                        | 已完成                | `g9ku2-web-404-fallback-theme-parity/SPEC.md`                  | 2026-04-07 | unknown HTML-route fallback now follows shared theme storage/system resolution, with Storybook light/dark proof, live `/accounts` dark proof, and cascade regression coverage                                                                                                                                    |
