+++ src/server/state.rs
@@ -20,6 +20,7 @@ struct DashboardOverviewFreshness {
     summary_last_activity: Option<i64>,
     summary_window_starts: [i64; 3],
     dashboard_rollup_signature: [i64; 4],
+    pending_dashboard_rollup_signature: [i64; 10],
     dashboard_api_key_lifecycle_signature: [i64; 3],
     dashboard_quarantine_lifecycle_signature: [i64; 3],
     dashboard_exhausted_lifecycle_signature: [i64; 3],
